feat: Enhance navigation and activity details
Adds new navigation links for 'trip-planner' and 'sos' to the main header, improving app discoverability.

Refactors the `DiscoverFeedComponent` and `ActivityDetailComponent` to streamline imports and update styling for better visual appeal and user experience. Updates the main app title in `AppComponent` to "Miúdos a Bordo".

Introduces internationalization support for the login page by integrating the `L10nPipe` and updating the login form structure for better alignment with common patterns.

Removes the `CurrencyPipe` from `DiscoverFeedComponent` as it was not being used.